Separate mapped post-processing code into separate file
